♻️ refactor : 자체 회원가입 UseCase 폴더 이동 및 개인정보 수정 반환 타입 수정 - user -> auth - void -> user
